requirements.txt
flask==3.1.1
openai==1.82.0
openpyxl==3.1.5
python-dotenv==1.1.0
werkzeug==3.1.3
gspread==6.1.4
google-auth==2.38.0
Pillow==11.2.1
pywebview==5.3.2
pyinstaller==6.13.0
